pyproject.toml
[project]
name = "syncfastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    "jinja2>=3.1.6",
]
